Revert "Working sample PSQL container test"
This reverts commit 5dc37086be13618d7082e47176399cfe02c3af58.
