+++ modules/r2dbc/src/main/java/org/testcontainers/containers/R2dbcConnectionParams.java
@@ -1,0 +1,4 @@
+package org.testcontainers.containers;
+
+public class R2dbcConnectionParams {
+}
