+++ modules/r2dbc/src/main/java/org/testcontainers/containers/R2dbcDatabaseContainerProvider.java
@@ -1,0 +1,5 @@
+package org.testcontainers.containers;
+
+public class R2dbcDatabaseContainerProvider {
+
+}
