--- modules/r2dbc/src/main/java/org/testcontainers/r2dbc/R2dbcConnectionParams.java
@@ -1,14 +1,0 @@
-package org.testcontainers.r2dbc;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class R2dbcConnectionParams {
-    private final String host;
-    private final int port;
-    private final String database;
-    private final String username;
-    private final String password;
-}
